+++ hololight/plotter.py
@@ -1,47 +1,94 @@
-import matplotlib.pyplot as plt
-import matplotlib.animation as animation
+from dataclasses import field
+
+import panel
+import param
+import ezmsg.core as ez
+
+import numpy as np
 
 from ezmsg.eeg.eegmessage import EEGMessage
 
-import ezmsg as ez
+import matplotlib
+matplotlib.use( "agg" )
 
-from typing import (
-    Optional
-)
+import matplotlib.pyplot as plt
+from matplotlib.figure import Figure
+from matplotlib.axes import Axes
+from matplotlib.lines import Line2D
 
-class PlotterSettings( ez.Settings ):
-    rate: float = 10.0 # Hz
+from typing import Optional, List
 
-class PlotterState( ez.State ):
-    data: Optional[ EEGMessage ] = None
+class EEGPlot( param.Parameterized ):
+    # Reactive Parameters
+    gain = param.Number( default = 1e6 )
+    offset = param.Number( default = 2.0 )
+    msg: Optional[ EEGMessage ] = param.Parameter()
 
-class Plotter( ez.Unit ):
+    # Internal plotting functionality
+    fig: Figure
+    ax: Axes
+    lines: Optional[ List[ Line2D ] ] = None
 
-    SETTINGS: PlotterSettings
-    STATE: PlotterState
+    def __init__( self, **params ) -> None:
+        super().__init__( **params )
+        self.fig, self.ax = plt.subplots( figsize = ( 8.0, 4.0 ) )
 
-    INPUT = ez.InputStream( EEGMessage )
+    def view( self ) -> Figure:
+        if self.msg is not None:
+            y_offsets = np.arange( self.msg.n_ch ) * self.offset
+            arr = ( self.msg.data * self.gain ) + y_offsets
+            eeg_dur = self.msg.n_time / self.msg.fs
+
+            if self.lines is None:
+                time = ( np.arange( self.msg.n_time ) / self.msg.fs ) - eeg_dur
+                self.lines = self.ax.plot( time, arr )
+            else:
+                self.ax.set_autoscale_on( False )
+                for idx, line in enumerate( self.lines ):
+                    line.set_ydata( arr[ :, idx ] )
+
+            self.ax.set_ylim( -self.offset, self.msg.n_ch * self.offset )
+            self.ax.set_yticks( y_offsets )
+            if self.msg.ch_names is not None:
+                self.ax.set_yticklabels( self.msg.ch_names )
+            self.ax.set_xlim( -eeg_dur, 0.0 )
+            self.ax.grid( 'True' )
+
+            self.ax.spines[ 'right' ].set_visible( False )
+            self.ax.spines[ 'top' ].set_visible( False )
+
+            self.fig.canvas.draw()
+
+
+        return self.fig
+
+    def panel( self ) -> panel.viewable.Viewable:
+        return panel.Row(
+            self.view,
+            panel.Column(
+                "<br>\n# EEG Plot",
+                panel.widgets.NumberInput.from_param( self.param[ 'gain' ] ),
+                panel.widgets.NumberInput.from_param( self.param[ 'offset' ] )
+            )
+        )
+
+class EEGPlotterSettings( ez.Settings ):
+    ...
+
+class EEGPlotterState( ez.State ):
+    plot: EEGPlot = field( default_factory = EEGPlot )
+
+class EEGPlotter( ez.Unit ):
+
+    SETTINGS: EEGPlotterSettings
+    STATE: EEGPlotterState
+
+    INPUT_SIGNAL = ez.InputStream( EEGMessage )
+
+    @ez.subscriber( INPUT_SIGNAL )
+    async def on_signal( self, msg: EEGMessage ) -> None:
+        self.STATE.plot.msg = msg
 
     @ez.main
-    def spawn_plot( self ) -> None:
-        # Matplotlib requires blocking on the main thread... :( 
-        # TODO: Consider moving self.ax to STATE.
-        fig, self.ax = plt.subplots( dpi = 100, figsize = ( 4.0, 4.0 ) )
-        anim = animation.FuncAnimation(
-            fig, self._animate, interval = 1 / self.SETTINGS.rate * 1000
-        )
-        plt.show()
-
-    def _animate( self, frame_idx: int ) -> None:
-        if self.ax is None or self.STATE.data is None:
-            return
-
-        self.ax.clear()
-        # self.ax.plot( ( self.STATE.data.data.T - self.STATE.data.data.mean( axis = 1 ) ).T )
-        self.ax.plot( self.STATE.data.data )
-
-        # self.ax.set_ylim( -2**23, 2**23 )
-
-    @ez.subscriber( INPUT )
-    async def plot( self, msg: EEGMessage ) -> None:
-        self.STATE.data = msg
+    def serve_dashboard( self ) -> None:
+        panel.serve( self.STATE.plot.panel, port = 8082 )
